refactor: extract common key helpers
