+++ x/shared/types/keys.go
@@ -1,5 +1,7 @@
 package types
 
+import "encoding/binary"
+
 const (
 	// ModuleName defines the module name
 	ModuleName = "shared"
@@ -17,4 +19,29 @@ var (
 
 func KeyPrefix(p string) []byte {
 	return []byte(p)
+}
+
+// IntKey converts an integer value to a byte slice for use in store keys
+// Appends a '/' separator to the end of the key for consistent prefix scanning
+func IntKey(intIndex int64) []byte {
+	var key []byte
+
+	heightBz := make([]byte, 8)
+	binary.BigEndian.PutUint64(heightBz, uint64(intIndex))
+	key = append(key, heightBz...)
+	key = append(key, []byte("/")...)
+
+	return key
+}
+
+// StringKey converts a string value to a byte slice for use in store keys
+// Appends a '/' separator to the end of the key for consistent prefix scanning
+func StringKey(strIndex string) []byte {
+	var key []byte
+
+	strIndexBz := []byte(strIndex)
+	key = append(key, strIndexBz...)
+	key = append(key, []byte("/")...)
+
+	return key
 }
